Fixed a small MySQL/Mariadb item
The question to remove rpm's from the mysql directory was appearing at
every step during a multi-step mysql upgrade. Fixed it so that this
question is only shown once.

Also fixed missing Mariadb upgrade check so that trying to 'upgrade' to
a lower version and same version is no longer possible.