Small bug in php upgrade fixed
Bash option -f unexpectedly also returns OK if no value is given,
causing the script to choose the wrong if/else route if var ALT_INI_FILE
is empty. Fixed this by populating this var with a non-existing
filename: "thisfiledoesnotexist01"

Also added a check for the "build" file in the rollback function, and if
found replace the original one with this file. This only happens if the
last action taken was a php upgrade to 5.6 in custombuild 1.x.